Inestable

+++ Kernel/libs/include/list.h
@@ -20,7 +20,7 @@ typedef struct PCB
 {
     int PID;
     int PPID;
-    uintptr_t RSP;
+    uint64_t RSP;
     char state;
     char priority;
     int FD[256];
